requirements.txt
streamlit==1.41.1
tensorflow==2.20.0
numpy==1.26.4
opencv-python==4.11.0.86
Pillow==11.2.0
scikit-learn==1.6.0
matplotlib==3.10.1
boto3==1.35.67
requests==2.32.3
python-dotenv==1.0.1
